% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Laplacian.R
\name{Laplacian}
\alias{Laplacian}
\title{Calculate the graph Laplacian from a given adjacency matrix}
\usage{
Laplacian(
  A,
  laplacian = c("sym", "rw"),
  grf.type = c("full", "knn", "e-graph"),
  k = 5,
  rho = NULL,
  epsilon = 0,
  mutual = FALSE,
  binary.grf = FALSE,
  plots = TRUE
)
}
\arguments{
\item{A}{An n by n kernel matrix, where n is the sample size, that represents your initial adjacency matrix. Kernel matrices are symmetric, positive semi-definite distance matrices}

\item{laplacian}{Either \code{"rw"} for the random walk laplacian (default), i.e., \eqn{I-D^{-1}A}, or \code{"sym"} for a symmetric Laplacian, i.e., \eqn{I-D^{-1/2}AD^{1/2}}}

\item{grf.type}{Type of graph to calculate: \code{"full"} for adjacency matrix equal to the kernel, \code{"knn"} for a k-nearest neighbors graph, \code{"e-graph"} for an "epsilon graph"}

\item{k}{An integer value for \code{k} in the k-nearest neighbors graph. Only the \code{k} largest edges (most similar neighbors) will be kept}

\item{rho}{A value for the dispersion parameter in the Gaussian kernel. It is in the denominator of the exponent, so higher values correspond to lower similarity. By default it is the median pairwise Gaussian distance}

\item{epsilon}{The cutoff value for the \code{e-graph}. Edges lower than this value will be removed}

\item{mutual}{Make a "mutual" knn graph. Only keeps edges when two nodes are both in each others k-nearest set}

\item{binary.grf}{Set all edges >0 to 1}

\item{plots}{Whether or not to plot the final graph, a heatmap of calculated kernel, and the eigen values of the Laplacian}
}
\value{
An n\eqn{\times}n matrix where \code{n} is the number of rows in \code{dat}.
}
\description{
Calculate the graph laplacian from a given kernel matrix that represents the full graph weighted adjacency matrix
}
\examples{

## Generating data with 3 distinct clusters
## Note that 'clustStruct' returns a list
dd <- clustStruct(n=120, p=30, k=3, noiseDat='random')[[1]]

## Gaussian kernel
rho <- median(dist(dd))
A <- exp(-(1/rho)*as.matrix(dist(dd, method = "euclidean", upper = TRUE)^2))

Laplacian(A, laplacian='rw', grf.type = 'knn')
}
\references{
\url{https://academic.oup.com/bioinformatics/article/36/4/1159/5566508#199177546}
}
