% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trueGroups.R
\name{trueGroups}
\alias{trueGroups}
\title{Get the groups created by the clustStruct function}
\usage{
trueGroups(n, k)
}
\arguments{
\item{n}{An integer, the sample size for all generated data sets}

\item{k}{An integer or vector, the number of distinct clusters in each generated data set. \code{n/k} must be an integer for all values of \code{k}}
}
\value{
A matrix with the unique groups/clusters from the multi-view data generated from \link[MMOC]{clustStruct}. The final column \code{Grps} enumerates these groups.
}
\description{
Get the unique groups generated by the \link[MMOC]{clustStruct} function for a given \code{k}. The number of rows of the resulting matrix gives the number of unique groups.
}
\examples{

trueGroups(n=120, k=c(2,3,4))

}
